chore: rm_extra_src

+++ src/printint.c
@@ -6,11 +6,12 @@
 /*   By: ktakada <ktakada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/30 15:42:37 by ktakada           #+#    #+#             */
-/*   Updated: 2022/05/30 15:48:52 by ktakada          ###   ########.fr       */
+/*   Updated: 2022/05/30 17:03:46 by ktakada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
+/* #include <stdio.h> */
 
 static int	get_digit_count(int num);
 
@@ -30,13 +31,17 @@ int	printint(int num)
 		print_count += write (1, "-", 1);
 	while (digit_count > 0)
 	{
+		/* printf("\ndigit_count: %d\n", digit_count); */
 		digit_count--;
-		print_num = num / recursive_power(10, digit_count);
+		/* printf("\npower: %d\n", (int)recursive_power(10, digit_count)); */
+		print_num = num / (int)recursive_power(10, digit_count);
 		if (neg)
 			print_num *= -1;
+		/* printf("\nprint_num: %d\n", print_num); */
 		print_num += '0';
 		print_count += write(1, &print_num, 1);
-		num %= recursive_power(10, digit_count);
+		num %= (int)recursive_power(10, digit_count);
+		/* printf("\n%d\n", num); */
 	}
 	return (print_count);
 }
@@ -55,3 +60,8 @@ int	get_digit_count(int num)
 	}
 	return (digit_count);
 }
+
+/* int	main(void) */
+/* { */
+/* 	printint(-10); */
+/* } */
